+++ resourcet-extra/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Version 0.1.1
 
-* Add `acquire1`, `acquireType1`.
+* Add `acquire1`, `acquireType1`, `unAcquire`.
 
 * Safer `mkAcquire1` and `mkAcquireType1` internals.